Include non-public metadata spatial extents containing bounding polygons in pdf documents if the user requesting the pdf has access to view the metadata record
